requirements.txt
# Core dependencies
aiohttp>=3.9.0
pyyaml>=6.0
python-dateutil>=2.8.0

# Optional: Database support
# psycopg2-binary>=2.9.0  # PostgreSQL
# sqlalchemy>=2.0.0

# Development dependencies (install with pip install -e ".[dev]")
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.0.0
# ruff>=0.1.0
# mypy>=1.0.0
